+++ src/Services/Telephony/Requests/Events/Auth.php
@@ -39,6 +39,3 @@ class Auth extends AbstractItem
         };
     }
 }
-
-
-