+++ src/Services/Telephony/Result/ExternalCallRegisterItemResult.php
@@ -16,14 +16,22 @@ namespace Bitrix24\SDK\Services\Telephony\Result;
 use Bitrix24\SDK\Core\Result\AbstractItem;
 
 /**
+ * If registration of the call was unsuccessful, the LEAD_CREATION_ERROR field will contain the error message.
+ *
  * @property-read  string $CALL_ID
- * @property-read  int $CRM_CREATED_LEAD
- * @property-read  int $CRM_ENTITY_ID
+ * @property-read  ?int $CRM_CREATED_LEAD
+ * @property-read  ?int $CRM_ENTITY_ID
  * @property-read  string $CRM_ENTITY_TYPE
  * @property-read  array $CRM_CREATED_ENTITIES
  * @property-read  string $LEAD_CREATION_ERROR
  */
 class ExternalCallRegisterItemResult extends AbstractItem
 {
-
+    /**
+     * @return bool
+     */
+    public function isError(): bool
+    {
+        return $this->data['LEAD_CREATION_ERROR'] !== '' && $this->data['LEAD_CREATION_ERROR'] !== null;
+    }
 }