+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 * add `Symfony\Component\Uid\Uuid` requirements
 * add contracts for bitrix24 applications based on bitrix24-php-sdk  - `Bitrix24\SDK\Application\Contracts`, now added `Bitrix24Account`
+* add [service builder factory](https://github.com/mesilov/bitrix24-php-sdk/issues/328)
 
 ### Changed
 * ❗️Batch interface `BatchInterface` [renamed](https://github.com/mesilov/bitrix24-php-sdk/issues/324) to `Bitrix24\SDK\Core\Contracts\BatchOperationsInterface`